Pt_Sleep now working

+++ portmidi/trunk/porttime/ptlinux.c
@@ -124,7 +124,7 @@ PtTimestamp Pt_Time()
 }
 
 
-void Pt_Sleep(int duration)
+void Pt_Sleep(long duration)
 {
     usleep(duration * 1000);
 }
